backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
apscheduler==3.10.4
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
# Optional providers
# tinyllm
google-api-python-client==2.146.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
# icalendar
